It's like s/long/int32/ only you don't end up with comments like "this doesn't beint32 here"
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@3909 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/csalerts.h
@@ -52,11 +52,11 @@ extern void vhalt(
 
 extern void _alephone_assert(
 	const char *file,
-	long line,
+	int32 line,
 	const char *what) NORETURN;
 extern void _alephone_warn(
 	const char *file,
-	long line,
+	int32 line,
 	const char *what);
 
 #if TARGET_API_MAC_CARBON && !defined(SDL)
