+++ Source_Files/CSeries/timer.h
@@ -10,19 +10,19 @@ become very unstable if your app crashes.
 
 */
 
-	extern unsigned long globalTime;
+	extern uint32 globalTime;
 
 #define MSECS_PER_CLICK 1
 
 class Timer {
-unsigned long totalTime;
-unsigned long startTime;
+uint32 totalTime;
+uint32 startTime;
 public:
 	Timer() {Clear();}
 	void Clear() { totalTime=0; startTime=0;}
 	void Start() { startTime=globalTime; }
 	void Stop() { totalTime += globalTime-startTime; startTime=0;}
-	unsigned long Clicks();
+	uint32 Clicks();
 	float Seconds() {return Clicks()/(1000.0/MSECS_PER_CLICK);}
 };
