+++ Source_Files/Files/extensions.h
@@ -42,7 +42,7 @@ void set_to_default_physics_file(void);
 /* Proceses the entire physics file.. */
 void import_definition_structures(void);
 
-void *get_network_physics_buffer(long *physics_length);
+void *get_network_physics_buffer(int32 *physics_length);
 void process_network_physics_model(void *data);
 
 #endif
